refactor(report): extract logger classes into report_logger.py
Issue #46 (EPIC-07-ST-02). ReportLogger (269 LOC) und ReportConsoleLogger
(80 LOC) aus services/report_agent.py in eigenes Modul
backend/app/services/report_logger.py verschoben.

- Re-Export im Service-Modul hält Modul-interne Aufrufstellen und
  potentielle Test-Mocks stabil. Externe Caller existieren keine.
- Type-Hint-Cleanup beim Move: section_title/section_index als
  Optional[str]/Optional[int] statt str = None/int = None.
- import logging von zwei lokalen Imports in _setup_file_handler/close
  auf Modul-Ebene gehoben.
- report_agent.py schrumpft 3053 → 2705 LOC (-11 %); seit Pfad A-Beginn
  -15 % insgesamt.

Verifikation: npm run check 5/5 grün — Backend 517 passed, Frontend 40
passed, Lint clean, Build ok.

Slice A5 (Bonus-Slice nach Pfad A) von v0.9.0.

Closes #46
