+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Format angelehnt an [Keep a Changelog](https://keepachangelog.com/de/1.1.0/), Ve
 
 ### Geändert
 
+- **Issue #46 (EPIC-07-ST-02: Report-Logging trennen) abgeschlossen.** `ReportLogger` und `ReportConsoleLogger` (zusammen 352 LOC) aus `backend/app/services/report_agent.py` (3053 LOC) in neue Datei `backend/app/services/report_logger.py` (377 LOC) verschoben. Re-Export im Service-Modul hält Modul-interne Aufrufe und potentielle Test-Mocks stabil. Type-Hint-Cleanup beim Move (`Optional[str]`/`Optional[int]` statt `str = None`/`int = None`); `import logging` von zwei lokalen Imports auf Modul-Ebene gehoben. `report_agent.py` schrumpft auf **2705 LOC** — von 3184 LOC vor Pfad A nun −15 %. Verhalten unverändert, 517 Backend- + 40 Frontend-Tests grün. Arbeitsprotokoll: `docu/2026-05-01-issue46-report-logger-arbeitsprotokoll.md`.
 - **Issue #49 (EPIC-07-ST-05: Evidence-Layer explizit modellieren, p2) als bereits erledigt geschlossen.** Inventur nach Pfad A3 + Issue #103 zeigt: `EvidenceItem` und `ReportClaim` (Tracking pro Aussage inkl. `tool_name`/`query`/`agent_log_ref`) sind seit #45 in `backend/app/models/report.py`; `evidence_binder.py` (S4a) macht Cosine-basiertes Binding pro Claim mit Threshold; `confidence_calculator.py` (S6) implementiert die Vier-Komponenten-Formel als Vorbereitung für EPIC-15; Self-Evidence ist seit S5 in `audit_trail` ausgegliedert. Tests: `test_evidence_binder.py`, `test_confidence_calculator.py`, `test_report_manager.py`. Status-Doku: `docu/2026-05-01-issue49-evidence-layer-status.md`.
 - **Issue #45 (EPIC-07-ST-01: Report-Models extrahieren) abgeschlossen.** Sechs Datenklassen (`ReportStatus`, `ReportSection`, `ReportOutline`, `Report`, `EvidenceItem`, `ReportClaim`) aus `backend/app/services/report_agent.py` (3184 LOC) in neue Datei `backend/app/models/report.py` (165 LOC) verschoben. Re-Export im Service-Modul hält bestehende Caller (`api/report.py`, `api/runs.py`, `test_report_manager.py`, `test_report_export.py`) unverändert. Unused Imports `dataclasses.dataclass` und `enum.Enum` aus `report_agent.py` entfernt; Datei schrumpft auf 3053 LOC (−4 %). `models/__init__.py` exportiert die neuen Klassen mit. Verhalten unverändert, 517 Backend- + 40 Frontend-Tests grün. Arbeitsprotokoll: `docu/2026-05-01-issue45-report-models-arbeitsprotokoll.md`.
 - **Issue #44 (EPIC-06-ST-04: Branching-Service extrahieren) abgeschlossen.** Neue Datei `backend/app/services/branching_service.py` (230 LOC) übernimmt `list_branches`, `create_branch` und `_apply_persona_overrides` als Funktionsmodul mit `SimulationManager` als Parameter — vermeidet zirkuläre Importe und hält die Manager-API stabil (keine Caller-Anpassungen in `api/simulation_profiles.py` oder Tests nötig). `simulation_manager.py` reduziert sich um 167 LOC auf 622 (−21 %); Manager-Methoden sind 1-Zeilen-Delegationen, unused Imports `shutil`/`Config`/`ArtifactLocator`/`RunRegistry` entfernt. `SimulationStatus`-Import in `create_branch` ist lazy, damit der Manager-→-Service-Import nicht in einen Zyklus kippt. Verhalten unverändert, 517 Backend-Tests + 40 Frontend-Tests grün. Arbeitsprotokoll: `docu/2026-05-01-issue44-branching-service-arbeitsprotokoll.md`.
